Setup ghost progress bar and fixed attribute clamping
7.63 - 7.64
Finished setting up a whitish globe behind the normal health/mana bar that interpolates to the value slower to create an afterimage of the change.

Clamp health / mana in PostGameplayEffectExecute so as to affect the actual value rather than just the modifier. This fixed a bug with the actual value going above the intended ceiling.